# openapi-atspaces.yaml
openapi: 3.0.3
info:
  title: At Spaces API
  version: 1.0.0
  description: |
    API contract for At Spaces (NestJS + Prisma + PostgreSQL + Redis).
    Roles: customer, vendor, admin. Access token in JSON, refresh token in HttpOnly cookie.
servers:
  - url: http://localhost:3000/api
  - url: https://<domain>/api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          items: { type: object }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        hasNext: { type: boolean }
...
    BranchListItem:
      type: object
      properties:
        id: { type: integer, example: 10 }
        name: { type: string, example: "Branch A" }
        city: { type: string, example: "Amman" }
        address: { type: string, example: "..." }
    FacilityPublic:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: "WiFi" }
        icon: { type: string, example: "wifi" }
        isAvailable: { type: boolean, example: true }
        description: { type: string, nullable: true }
    VendorServiceInBranch:
      type: object
      properties:
        vendorServiceId: { type: integer, example: 55 }
        serviceId: { type: integer, example: 1 }
        name: { type: string, example: "Hot Desk" }
        pricePerUnit: { type: number, format: float, example: 5.0 }
        priceUnit: { type: string, enum: [hour, day, week, month], example: hour }
        maxCapacity: { type: integer, example: 50 }
        isAvailable: { type: boolean, example: true }
    BranchDetails:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        city: { type: string }
        address: { type: string }
        latitude: { type: number, format: float, nullable: true }
        longitude: { type: number, format: float, nullable: true }
        facilities:
          type: array
          items: { $ref: '#/components/schemas/FacilityPublic' }
        services:
          type: array
          items: { $ref: '#/components/schemas/VendorServiceInBranch' }
    AvailabilityCheckRequest:
      type: object
      required: [vendorServiceId, startTime, endTime, quantity]
      properties:
        vendorServiceId: { type: integer, example: 55 }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        quantity: { type: integer, minimum: 1, example: 2 }
...
    BookingPreviewResponse:
      type: object
      properties:
        totalPrice: { type: number, format: float, example: 20.0 }
        currency: { type: string, example: JOD }
    CreateBookingRequest:
      allOf:
        - $ref: '#/components/schemas/AvailabilityCheckRequest'
        - type: object
          required: [paymentMethod]
          properties:
            paymentMethod: { type: string, enum: [cash, card, apple_pay] }
    CreateBookingResponse:
      type: object
      properties:
        bookingId: { type: integer, example: 900 }
        bookingNumber: { type: string, example: "BKG-20260305-0001" }
        totalPrice: { type: number, format: float, example: 20.0 }
        status: { type: string, example: pending }
        paymentStatus: { type: string, example: pending }
    BookingListItem:
      type: object
      properties:
        id: { type: integer }
        bookingNumber: { type: string }
        branchName: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        status: { type: string }
    BookingDetails:
      allOf:
        - $ref: '#/components/schemas/BookingListItem'
        - type: object
          properties:
            quantity: { type: integer }
            totalPrice: { type: number, format: float }
            vendorServiceId: { type: integer }
...
    NotificationItem:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, example: booking_confirmed }
        title: { type: string }
        body: { type: string }
        isRead: { type: boolean, example: false }
        createdAt: { type: string, format: date-time }
...
paths:
...
  /branches:
    get:
      tags: [Branches]
      summary: List branches (filters + pagination)
      parameters:
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: serviceId
          in: query
          required: false
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged branches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/BranchListItem' }
  /branches/search:
    get:
      tags: [Branches]
      summary: Search branches by query string
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged branches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/BranchListItem' }
  /branches/{id}:
    get:
      tags: [Branches]
      summary: Branch details (facilities + services)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Branch details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchDetails' }
...
  /bookings/preview:
    post:
      tags: [Bookings]
      summary: Preview booking price (no booking created)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvailabilityCheckRequest' }
      responses:
        '200':
          description: Price preview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingPreviewResponse' }
  /bookings:
    post:
      tags: [Bookings]
      summary: Create booking
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingRequest' }
      responses:
        '201':
          description: Created booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateBookingResponse' }
  /bookings/my:
    get:
      tags: [Bookings]
      summary: List my bookings
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: My bookings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/BookingListItem' }
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Booking details (ownership required)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingDetails' }
  /bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel booking
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /bookings/{id}/calendar.ics:
    get:
      tags: [Bookings]
      summary: Export booking to calendar (ICS)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: ICS calendar file
          content:
            text/calendar:
              schema:
                type: string
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/NotificationItem' }
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark notification as read
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
...